requirements.txt
keras==2.12.0
numpy==1.23.5
numpy==1.25.0
opencv_python==4.7.0.72
opencv_python_headless==4.8.0.74
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.24.0
